+++ Etomo/src/etomo/comscript/Utilities.java
@@ -22,6 +22,10 @@ import etomo.util.Montagesize;
  * @version $Revision$
  * 
  * <p> $Log$
+ * <p> Revision 1.3  2010/02/17 04:47:54  sueh
+ * <p> bug# 1301 Using the manager instead of the manager key do pop up
+ * <p> messages.
+ * <p>
  * <p> Revision 1.2  2009/03/17 00:33:25  sueh
  * <p> bug# 1186 Pass managerKey to everything that pops up a dialog.
  * <p>
@@ -33,7 +37,7 @@ import etomo.util.Montagesize;
 final class Utilities {
   public static final String rcsid = "$Id$";
 
-  static boolean is90DegreeImageRotation(float imageRotation) {
+  static boolean is90DegreeImageRotation(double imageRotation) {
     return (imageRotation > 45 && imageRotation < 135)
         || (imageRotation < -45 && imageRotation > -135);
   }
